fix: address 5 automated review comments on bridge selectors
